ShowObject: Don't follow circular references
To identify circular references it now records all reference types it
encounters.  Whenever it encounters a new reference type it checks to
see if there is already a reference to the same object in its records.
If there is it will not display the same object twice but display a
message that it has encountered a circular reference.

Also add a ShowObject overload with new parameter simpleDataTypesOnly,
which will only display members which are value types or strings, not
reference types.